Work around due to gatsby cloud not supporting ESM . Readme fix
