+++ gatsby/.env.example
@@ -18,5 +18,3 @@ MAIL_USER = "${MAIL_USER}"
 MAIL_PASS = "${MAIL_PASS}"
 
 NETLIFY_USE_YARN = "${MAIL_PASS}"
-
-NODE_OPTIONS = "${NODE_OPTIONS}"
